First build
This is pretty much just an unfinished version of the playmate framework I've been building. The only things configurable right now are the body, head, eyes, nose, and skin. There's no UI, practically no interaction, it's just the basics of the framework and just enough for me to test things out on hardware.

First of many commits hopefully!